refactor: reorganize folders and update imports to adopt Screaming Architecture
