+++ frontend/src/app/layouts/Footer/Footer.tsx
@@ -1,5 +1,5 @@
 import { Link } from "react-router";
-import Button from "./Button";
+import Button from "@/app/ui/Button";
 import tikTok from "@/assets/tikTok.webp";
 import instagram from "@/assets/instagram.webp";
 import youtube from "@/assets/youtube.webp";
