+++ frontend/src/auth/components/LoginForm.tsx
@@ -1,8 +1,8 @@
 import { zodResolver } from "@hookform/resolvers/zod";
 import { SubmitHandler, useForm } from "react-hook-form";
 import { z } from "zod";
-import Button from "@/shared/components/Button";
-import Input from "@/shared/components/Input";
+import Button from "@/app/ui/Button";
+import Input from "@/app/ui/Input";
 import { EyeOff } from "./ui/EyeOff";
 import { EyeOn } from "./ui/EyeOn";
 import { useState } from "react";
