+++ frontend/src/auth/components/RoleSelector.tsx
@@ -1,5 +1,5 @@
 import image from "@/assets/image.webp";
-import Button from "@/shared/components/Button";
+import Button from "@/app/ui/Button";
 
 const ROLES = ["FAN", "MUSICIAN", "ORGANIZER"];
