+++ frontend/src/routes/index.tsx
@@ -1,6 +1,6 @@
 import { Route, Routes } from "react-router";
 import { HomePage } from "@/home/components/HomePage";
-import { MainLayout } from "@/shared/components/layouts/MainLayout";
+import { MainLayout } from "@/app/layouts/MainLayout";
 import ProfilePage from "@/profiles/components/ProfilePage";
 import ExplorerPage from "@/explorer/components/ExplorerPage";
 import ArtistPage from "@/artist/components/ArtistPage";
