+++ frontend/src/shared/components/Cards/StackedCard.tsx
@@ -1,4 +1,4 @@
-import Button from "../Button";
+import Button from "@/app/ui/Button";
 
 interface StackedCardProps {
   titleHeader: string;
