--- frontend/src/shared/components/apiAxios.tsx
@@ -1,8 +1,0 @@
-import axios from "axios";
-
-export const apiAxios = axios.create({
-  baseURL: "https://ecos-ed30.onrender.com",
-  headers: {
-    "Content-Type": "application/json",
-  },
-});
